+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to the "bang-log" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.2] - 2025-10-24
+
+### Fixed
+- Fixed console.log insertion position to properly insert on a new line instead of same line
+- Improved handling of document edge cases when inserting at the end of file
+
+### Added
+- Added test case for Hook calls with object arguments
+
 ## [Unreleased]
 
 - Initial release